requirements.txt
altgraph==0.17.3
anyio==3.6.2
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-generator==1.10
attrs==22.2.0
autopep8==2.0.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==6.0.0
bs4==0.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
comm==0.1.2
contourpy==1.0.7
cycler==0.11.0
darkdetect==0.7.1
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
et-xmlfile==1.1.0
executing==1.2.0
fastjsonschema==2.16.2
flake8==6.0.0
fonttools==4.38.0
fqdn==1.5.1
h11==0.14.0
idna==3.4
ipykernel==6.21.2
ipython==8.10.0
ipython-genutils==0.2.0
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
jsonpointer==2.3
jsonschema==4.17.3
jupyter-events==0.6.3
jupyter_client==8.0.2
jupyter_core==5.2.0
jupyter_server==2.2.1
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
kiwisolver==1.4.4
macholib==1.16.2
MarkupSafe==2.1.2
matplotlib==3.7.0
matplotlib-inline==0.1.6
mccabe==0.7.0
mistune==2.0.5
mypy==0.991
mypy-extensions==0.4.3
nbclassic==0.5.1
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
nest-asyncio==1.5.6
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.24.2
openpyxl==3.1.1
outcome==1.2.0
packaging==23.0
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pip-autoremove==0.10.0
platformdirs==3.0.0
prometheus-client==0.16.0
prompt-toolkit==3.0.36
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
Pygments==2.14.0
pyinstaller==5.8.0
pyinstaller-hooks-contrib==2023.0
pyparsing==3.0.9
PyPDF2==3.0.1
pyqtdarktheme==2.1.0
pyrsistent==0.19.3
PySide6==6.4.2
PySide6-Addons==6.4.2
PySide6-Essentials==6.4.2
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
python-json-logger==2.0.6
pytz==2022.6
PyYAML==6.0
pyzmq==25.0.0
requests==2.28.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
selenium==4.8.0
Send2Trash==1.8.0
shiboken6==6.4.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
stack-data==0.6.2
terminado==0.17.1
tinycss2==1.2.1
tomli==2.0.1
tornado==6.2
traitlets==5.9.0
trio==0.22.0
trio-websocket==0.9.2
types-python-dateutil==2.8.19.4
types-pytz==2022.6.0.1
types-requests==2.28.11.7
types-urllib3==1.26.25.4
typing_extensions==4.4.0
uri-template==1.2.0
urllib3==1.26.14
wcwidth==0.2.6
webcolors==1.12
webencodings==0.5.1
websocket-client==1.5.1
wsproto==1.2.0